json de relais en fonction du master

+++ relais-supervision/relais-loop.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
-relaisjson=relais.json
+if [ ! $# -eq 1 ]; then
+        echo "Usage: $0 <stream_name>"
+        exit 1
+fi
+stream_name=$1
+relaisjson=relais-$stream_name.json
+[ ! -f $relaisjson ] && cp relais.json $relaisjson
+
 rm /var/lock/.relayingstuff
 relais=$(cat $relaisjson)
 echo $relais |jq -c '.[]'|while read json
